refactored seeds and created review seeds file
