+++ db/seeds.rb
@@ -6,1199 +6,1205 @@
 #   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
-User.create({
-  first_name: "Guest",
-  last_name: "User",
-  email: "guest@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482131647/couchbnb/person-solid_telh7f.png"
-})
 
-User.create({
-  first_name: "Jay",
-  last_name: "Hwang",
-  email: "jay_hwang@gmail.com",
-  password: "password",
-  profile_pic_url: "http://res.cloudinary.com/ddgtwtbre/image/upload/v1481587215/portfolio/linkedin_eiulzq.jpg"
-})
-
-User.create({
-  first_name: "Adam",
-  last_name: "Foosaner",
-  email: 'adam@gmail.com',
-  password: 'password',
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Adam-Foosaner_nypijr.png"
-})
-
-User.create({
-  first_name: "Alex",
-  last_name: "Bennet",
-  email: "alex@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alex-Bennet_bpdgy6.png"
-})
-
-User.create({
-  first_name: "Alex",
-  last_name: "Manning",
-  email: "alex_manning@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alex-Manning_tzlryt.png"
-})
-
-User.create({
-  first_name: "Alex",
-  last_name: "Doundakov",
-  email: "alex_doundakov@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alexander-Doundakov_g045nm.png"
-})
-
-User.create({
-  first_name: "Amanda",
-  last_name: "Fielding",
-  email: "amanda@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128657/couchbnb/Amanda-Fielding_fnmudx.png"
-})
-
-User.create({
-  first_name: "Andrew",
-  last_name: "Tae",
-  email: "andrew_tae@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Andrew-Tae_ht1kcz.png"
-})
-
-User.create({
-  first_name: "Anudeep",
-  last_name: "Kasturi",
-  email: "anudeep_kasturi@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128664/couchbnb/Anudeep-Kasturi_odxgby.png"
-})
-
-User.create({
-  first_name: "Benjamin",
-  last_name: "Dippelsman",
-  email: "ben_dip@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128658/couchbnb/Benjamin-Dippelsman_xl0tbl.png"
-})
-
-User.create({
-  first_name: "Brian",
-  last_name: "Chen",
-  email: "brian_chen@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/Brian-Chen_mwcxww.png"
-})
-
-User.create({
-  first_name: "Brian",
-  last_name: "Lee",
-  email: "brian_lee@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128666/couchbnb/Brian-Lee_l4b2h5.png"
-})
-
-User.create({
-  first_name: "Brian",
-  last_name: "Tsai",
-  email: "brian_tsai@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Brian-Tsai_n5qaky.png"
-})
-
-User.create({
-  first_name: "Caleb",
-  last_name: "Ontiveros",
-  email: "caleb_onti@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Caleb-Ontiveros_ihuvpj.png"
-})
-
-User.create({
-  first_name: "Christopher",
-  last_name: "Button",
-  email: "christopher_button@gmail.com",
-  password: 'password',
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/Christopher-Button_o3lysh.png"
-})
-
-User.create({
-  first_name: "Cindy",
-  last_name: "Wang",
-  email: "cindy_wang@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/Cindy-Wang_woewj1.png"
-})
-
-User.create({
-  first_name: "Connor",
-  last_name: "Murphy",
-  email: "connor_muphy@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128668/couchbnb/Connor-Murphy_kmoitq.png"
-})
-
-User.create({
-  first_name: "Daniel",
-  last_name: "Sulfaro",
-  email: "daniel_sulfaro@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128667/couchbnb/Daniel-Sulfaro_lnsv0o.png"
-})
-
-User.create({
-  first_name: "David",
-  last_name: "Kim",
-  email: "david_kim@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/David-Kim_nqb3b7.png"
-})
-
-User.create({
-  first_name: "David",
-  last_name: "McAfee",
-  email: 'david_mcafee@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/David-McAfee_v2okis.png"
-})
-
-User.create({
-  first_name: "Debra",
-  last_name: "Fong",
-  email: "debra_fong@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128668/couchbnb/Debra-Fong_junc34.png"
-})
-
-User.create({
-  first_name: "Deep",
-  last_name: "Tailor",
-  email: "deep_tailor@gmail.com",
-  password: "password",
-  profile_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/Deep-Tailor_fcid09.png'
-})
-
-User.create({
-  first_name: "Duc",
-  last_name: "Nguyen",
-  email: "duc_nguyen@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128670/couchbnb/Duc-Nguyen_tqn8ll.png"
-})
-
-User.create({
-  first_name: "Eric",
-  last_name: "Kwok",
-  email: "eric_kwok@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/Eric-Kwok_tijkgu.png"
-})
-
-User.create({
-  first_name: "Eric",
-  last_name: "Oliveira",
-  email: 'eric_oliveira@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Eric-Oliveira_wpe6od.png"
-})
-
-User.create({
-  first_name: "Ethan",
-  last_name: "Wan",
-  email: "ethan_wan@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Ethan-Wan_ekbgwv.png"
-})
-
-User.create({
-  first_name: "Evan",
-  last_name: "Waits",
-  email: "evan_waits@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Ethan-Wan_ekbgwv.png"
-})
-
-User.create({
-  first_name: "Fan",
-  last_name: "Song",
-  email: "fan_song@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Fan-Song_ul7irn.png"
-})
-
-User.create({
-  first_name: "Frank",
-  last_name: "Bi",
-  email: 'frank_bi@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Frank-Bi_yczosf.png"
-})
-
-User.create({
-  first_name: "Gabriel",
-  last_name: "Lewis",
-  email: "gabriel_lewis@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128673/couchbnb/Gabriel-Lewis_srgc8n.png"
-})
-
-User.create({
-  first_name: "Humza",
-  last_name: "Baig",
-  email: 'humza_baig@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Hamza-Baig_jnvfo1.png"
-})
-
-User.create({
-  first_name: "Henry",
-  last_name: "Hsu",
-  email: "henry_hsu@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Hamza-Baig_jnvfo1.png"
-})
-
-User.create({
-  first_name: "James",
-  last_name: "Conklin",
-  email: "james_conklin",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128678/couchbnb/James-Conklin_p0dvzr.png"
-})
-
-User.create({
-  first_name: "Jane",
-  last_name: "Hauf",
-  email: "jane_hauf@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Jane-Hauf_chq1eb.png"
-})
-
-User.create({
-  first_name: "Janine",
-  last_name: "Love",
-  email: 'janine_love@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128676/couchbnb/Janine-Love_fjhlft.png"
-})
-
-User.create({
-  first_name: 'JC',
-  last_name: 'Balcita',
-  email: "jc_balcita@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128679/couchbnb/JC-Balcita_uhjc9l.png"
-})
-
-User.create({
-  first_name: "John",
-  last_name: "Lekili",
-  email: 'john_lekili@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128678/couchbnb/John-Lekili_hjymxe.png"
-})
-
-User.create({
-  first_name: "Joseph",
-  last_name: "McBride",
-  email: 'joseph_mcbride@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128681/couchbnb/Joseph-McBride_rt9bro.png"
-})
-
-User.create({
-  first_name: "Justice",
-  last_name: "Yen",
-  email: "justice_yen@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128679/couchbnb/Justice-Yen_vaqjbg.png"
-})
-
-User.create({
-  first_name: "Lindsey",
-  last_name: "Whitley",
-  email: 'lindsey_whitley@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Lindsey-Whitley_pwedp2.png"
-})
-
-User.create({
-  first_name: "Louis",
-  last_name: "Cruz",
-  email: "louis_cruz@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Louis-Cruz_bwx4za.png"
-})
-
-User.create({
-  first_name: "Luke",
-  last_name: "Wassink",
-  email: "luke_wassink@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Luke-Wassink_dvzgti.png"
-})
-
-User.create({
-  first_name: "Matt",
-  last_name: "Velker",
-  email: "matt_velker@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128680/couchbnb/Matt-Velker_snyprp.png"
-})
-
-User.create({
-  first_name: "Melissa",
-  last_name: "Lau",
-  email: "melissa_lau@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128683/couchbnb/Melissa-Lau_xrxrbh.png"
-})
-
-User.create({
-  first_name: "Micah",
-  last_name: "Sapitsky",
-  email: "micach_sapitsky@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128684/couchbnb/Micah-Sapitsky_yt84jv.png"
-})
-
-User.create({
-  first_name: "Michael",
-  last_name: "Hirabayashi",
-  email: "michael_hirabayashi@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Michael-Hirabayashi_jd05gl.png"
-})
-
-User.create({
-  first_name: "Michael",
-  last_name: "Loschiavo",
-  email: 'michael_loschiavo@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128688/couchbnb/Michael-Loschiavo_kzxqhl.png"
-})
-
-User.create({
-  first_name: "Michael",
-  last_name: "Pak",
-  email: "michael_pak@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Michael-Pak_dqyesg.png"
-})
-
-User.create({
-  first_name: "Michaelangelo",
-  last_name: "Loreti",
-  email: "miki_loreti@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Michaelangelo-Loreti_td9ell.png"
-})
-
-User.create({
-  first_name: 'Mike',
-  last_name: "Barille",
-  email: 'mike_barille@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128687/couchbnb/Mike-Barille_igrjta.png"
-})
-
-User.create({
-  first_name: "Miki",
-  last_name: "Sparks",
-  email: "miki_sparks@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128687/couchbnb/Miki-Sparks_yttcgo.png"
-})
-
-User.create({
-  first_name: "Nicholas",
-  last_name: "Skoularikis",
-  email: 'nicho@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Nicholas-Skoularikis_p71di2.png"
-})
-
-User.create({
-  first_name: "Paul",
-  last_name: "Cathey",
-  email: "paul_cathey@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128688/couchbnb/Paul-Cathey_uqbi6h.png"
-})
-
-User.create({
-  first_name: "Peik",
-  last_name: "Sia",
-  email: "peik_sia@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Peik-Sia_bqpfzr.png"
-})
-
-User.create({
-  first_name: "Pysith",
-  last_name: "Vanuptikul",
-  email: 'pi_van@gmail.com',
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Pysith_Vanuptikul_eh6roj.png"
-})
-
-User.create({
-  first_name: "Samuel",
-  last_name: "Mak",
-  email: "sam_mak@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128692/couchbnb/Samuel-Mak_e9huh3.png"
-})
-
-User.create({
-  first_name: "Sebastian",
-  last_name: "Cruz",
-  email: "seb_cruz@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128691/couchbnb/Sebastian-Cruz_lm5qde.png"
-})
-
-User.create({
-  first_name: "Shaurya",
-  last_name: "Sinha",
-  email: "shaurya_sinha@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128690/couchbnb/Shaurya-Sinha_fbwey8.png"
-})
-
-User.create({
-  first_name: "Sonik",
-  last_name: "Jang",
-  email: "sonik_jang@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Sonik-Jhang_ajrt0r.png"
-})
-
-User.create({
-  first_name: "Taylor",
-  last_name: "Chestnut",
-  email: "taylor_chestnut",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128691/couchbnb/Taylor-Chestnut_wkwtrx.png"
-})
-
-User.create({
-  first_name: "Todd",
-  last_name: "Nestor",
-  email: "todd_nestor@gmail.com",
-  password: "password",
-  profile_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128693/Todd-Nestor_ehdiai.png'
-})
-
-User.create({
-  first_name: "Tonia",
-  last_name: "Hsia",
-  email: "tonia_hsia@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Tonia-Hsia_dxh834.png"
-})
-
-User.create({
-  first_name: "Tony",
-  last_name: "Xie",
-  email: "tony_xie@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Tony-Xie_prap45.png"
-})
-
-User.create({
-  first_name: "Travis",
-  last_name: "Ludlum",
-  email: "travis_ludlum@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128693/couchbnb/Travis-Ludlum_dnjeqg.png"
-})
-
-User.create({
-  first_name: "Trevor",
-  last_name: "Scandalios",
-  email: "trevor_scan@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Trevor-Scandalios_lwkfma.png"
-})
-
-User.create({
-  first_name: "Tyler",
-  last_name: "Fields",
-  email: "tyler_fields@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Trevor-Scandalios_lwkfma.png"
-})
-
-User.create({
-  first_name: "Victor",
-  last_name: "Aw",
-  email: "victor_aw@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Victor-Aw_vesrpm.png"
-})
-
-User.create({
-  first_name: "Vince",
-  last_name: "Sangke",
-  email: "vince_sangke@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128695/couchbnb/Vince-Sangkhae_gotqet.png"
-})
-
-User.create({
-  first_name: "Winston",
-  last_name: "Zhao",
-  email: "winston_zhao@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Winston-Zhao_a5pqvw.png"
-})
-
-User.create({
-  first_name: "Yue",
-  last_name: "Gu",
-  email: "yue_gu@gmail.com",
-  password: "password",
-  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Yue-Gu_nxn82n.png"
-})
-
-Spot.create({
-  host_id: 2,
-  host_name: "Host",
-  title: "San Francisco Apt Downtown",
-  description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-  country: "United States",
-  state_region: "California",
-  city: "San Francisco",
-  post_code: '94103',
-  street_address: '12345 California St',
-  price_per_night: 100,
-  room_type: "Entire House/Apt",
-  bed_count: 4,
-  max_guests: 4,
-  spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic3_hdfvpl.jpg',
-  lat: 37.794144,
-  lng: -122.408957
-})
-
-Spot.create({
-  host_id: 2,
-  host_name: "Jay",
-  title: "Los Angeles Apt Downtown",
-  description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-  country: "United States",
-  state_region: "California",
-  city: "Los Angeles",
-  post_code: '91002',
-  street_address: '12345 Wilshire Blvd',
-  price_per_night: 100,
-  room_type: "Entire House/Apt",
-  bed_count: 4,
-  max_guests: 4,
-  spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic1_ojh6w2.jpg',
-  lat: 34.052256,
-  lng: -118.254381
-})
-
-
-Spot.create({
-  host_id: 2,
-  host_name: "Jay",
-  title: "Private Studio in London",
-  description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-  country: "United Kingdom",
-  state_region: "",
-  city: "London",
-  post_code: 'AG42NB',
-  street_address: '12345 California St',
-  price_per_night: 100,
-  room_type: "Private Room",
-  bed_count: 1,
-  max_guests: 2,
-  spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic4_r0oafw.jpg',
-  lat: 51.520915,
-  lng: -0.149832,
-})
-
-Spot.create({
-  host_id: 2,
-  host_name: "Jay",
-  title: "Quaint room in Los Angeles",
-  description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-  country: "United States",
-  state_region: "California",
-  city: "Los Angeles",
-  post_code: '91002',
-  street_address: '12345 Sunset Blvd',
-  price_per_night: 50,
-  room_type: "Shared Room",
-  bed_count: 2,
-  max_guests: 4,
-  spot_pic_url: 'http://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic2_ucamo9.jpg',
-  lat: 34.090932,
-  lng: -118.406058
-})
-
-Spot.create({
-  host_id: 2,
-  host_name: "Jay",
-  title: "Disneyland Paris perfect for 5!",
-  description: "This apartment is ideally located at the foot of the famous Valley Shopping and just close to the Disney Parks!! Finish a day full of emotions for a moment of relaxation in the pool of the residence.",
-  country: "France",
-  state_region: "Île-de-France",
-  city: "Serris",
-  post_code: "77700",
-  street_address: "12345 Paris Ave",
-  price_per_night: 75,
-  room_type: "Entire House/Apt",
-  bed_count: 2,
-  max_guests: 5,
-  spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/c_scale,h_604/v1478628133/paris-eiffel-tower_hxpxjx.jpg',
-  lat: 48.843499,
-  lng: 2.785134
-})
-
-Spot.create({
-  host_id: 2,
-  host_name: "Jay",
-  title: "Luxury Flat",
-  description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-  country: "Netherlands",
-  state_region: "Noord-Holland",
-  city: "Amsterdam",
-  post_code: "12010",
-  street_address: "12345 Amsterdam Street",
-  price_per_night: 147,
-  room_type: "Entire House/Apt",
-  bed_count: 6,
-  max_guests: 8,
-  spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478628600/amsterdam-luxury-flat_jkgxej.webp',
-  lat: 52.374172,
-  lng: 4.861534
-})
-
-Spot.create({
-  host_id: 2,
-  host_name: "Jay",
-  title: "Entertainer's Paradise, Heart of Umhlanga",
-  description: "Our contemporary home boasts open plan living at its finest. This fully furnished home is situated in a quiet and safe complex in Umhlanga, Durban. The house consists of four en-suite bedrooms, a fully equipped kitchen, 12 seating dining area & 2 TV lounges. There is a large infinity pool with heated Jacuzzi located off a sunset facing pool deck looking into natural bush.
-
-  A 2 minute drive to the beach, our home is the perfect escape.",
-  country: "South Africa",
-  state_region: 'KZN',
-  city: "Durban",
-  post_code: '12345',
-  street_address: "12345 Durban Poison Ave",
-  price_per_night: 801,
-  room_type: "Entire House/Apt",
-  bed_count: 5,
-  max_guests: 8,
-  spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478629001/durban-south-africa_yp7win.webp',
-  lat: -29.856882,
-  lng: 31.036159
-})
-
-Spot.create({
-  host_id: 2,
-  host_name: "Jay",
-  title: "Spacious Home in heart of Seoul",
-  description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-  country: "South Korea",
-  state_region: "",
-  city: "Seoul",
-  post_code: '09876',
-  street_address: "1234 Seoul Street",
-  price_per_night: 120,
-  room_type: "Entire House/Apt",
-  bed_count: 3,
-  max_guests: 6,
-  spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478629386/seoul-south-korea_a1lsjg.jpg',
-  lat: 37.513837,
-  lng: 127.029368
-})
-
-spot_locations = [
-  {
-    "city":"San Francisco",
-    "state_region":" California",
-    "country":" USA",
-    "lat":37.7609082,
-    "lng":-122.4350043
-  },
-
-  {
-    "city":"Los Angeles",
-    "state_region":" CA",
-    "country":" USA",
-    "lat":34.0522342,
-    "lng":-118.2436849
-  },
-
-  {
-    "city":"Hollywood",
-    "state_region":" Los Angeles",
-    "country":" USA",
-    "lat":34.0928092,
-    "lng":-118.3286614
-  },
-
-  {
-    "city":"Phoenix",
-    "state_region":" AZ",
-    "country":" USA",
-    "lat":33.4483771,
-    "lng":-112.0740373
-  },
-
-  {
-    "city":"Tucson",
-    "state_region":" AZ",
-    "country":" USA",
-    "lat":32.2217429,
-    "lng":-110.926479
-  },
-
-  {
-    "city":"El Paso",
-    "state_region":" TX",
-    "country":" USA",
-    "lat":31.7618778,
-    "lng":-106.4850217
-  },
-
-  {
-    "city":"Brooklyn",
-    "state_region":" NY",
-    "country":" USA",
-    "lat":40.6781784,
-    "lng":-73.9441579
-  },
-
-  {
-    "city":"Bronx County",
-    "state_region":" NY",
-    "country":" USA",
-    "lat":40.8370495,
-    "lng":-73.86542949999999
-  },
-
-  {
-    "city":"Queens County",
-    "state_region":" NY",
-    "country":" USA",
-    "lat":40.7282239,
-    "lng":-73.7948516
-  },
-
-  {
-    "city":"Manhattan",
-    "state_region":" New York",
-    "country":" USA",
-    "lat":40.7830603,
-    "lng":-73.9712488
-  },
-
-  {
-    "city":"Chicago",
-    "state_region":" IL",
-    "country":" USA",
-    "lat":41.8781136,
-    "lng":-87.6297982
-  },
-
-  {
-    "city":"Las Vegas",
-    "state_region":" NV",
-    "country":" USA",
-    "lat":36.1699412,
-    "lng":-115.1398296
-  },
-
-  {
-    "city":"Seattle",
-    "state_region":" WA",
-    "country":" USA",
-    "lat":47.6062095,
-    "lng":-122.3320708
-  },
-
-  {
-    "city":"Washington",
-    "state_region":" DC",
-    "country":" USA",
-    "lat":38.9071923,
-    "lng":-77.0368707
-  },
-
-  {
-    "city":"Miami",
-    "state_region":" FL",
-    "country":" USA",
-    "lat":25.7616798,
-    "lng":-80.1917902
-  },
-
-  {
-    "city":"Boston",
-    "state_region":" MA",
-    "country":" USA",
-    "lat":42.3600825,
-    "lng":-71.0588801
-  },
-
-  {
-    "city":"Portland",
-    "state_region":" OR",
-    "country":" USA",
-    "lat":45.5230622,
-    "lng":-122.6764816
-  },
-
-  {
-    "city":"Nashville",
-    "state_region":" TN",
-    "country":" USA",
-    "lat":36.1626638,
-    "lng":-86.7816016
-  },
-
-  {
-    "city":"Dallas",
-    "state_region":" TX",
-    "country":" USA",
-    "lat":32.7766642,
-    "lng":-96.79698789999999
-  },
-
-  {
-    "city":"Austin",
-    "state_region":" TX",
-    "country":" USA",
-    "lat":30.267153,
-    "lng":-97.7430608
-  },
-
-  {
-    "city":"Minneapolis",
-    "state_region":" MN",
-    "country":" USA",
-    "lat":44.977753,
-    "lng":-93.2650108
-  },
-
-  {
-    "city":"Seoul",
-    "state_region":"",
-    "country":" South Korea",
-    "lat":37.566535,
-    "lng":126.9779692
-  },
-
-  {
-    "city":"São Paulo",
-    "state_region":" State of São Paulo",
-    "country":" Brazil",
-    "lat":-23.5505199,
-    "lng":-46.63330939999999
-  },
-
-  {
-    "city":"Mumbai",
-    "state_region":" Maharashtra",
-    "country":" India",
-    "lat":19.0759837,
-    "lng":72.8776559
-  },
-
-  {
-    "city":"Moscow",
-    "state_region":"",
-    "country":" Russia",
-    "lat":55.755826,
-    "lng":37.6173
-  },
-
-  {
-    "city":"Shanghai",
-    "state_region":"",
-    "country":" China",
-    "lat":31.230416,
-    "lng":121.473701
-  },
-
-  {
-    "city":"Tokyo",
-    "state_region":"",
-    "country":" Japan",
-    "lat":35.7090259,
-    "lng":139.7319925
-  },
-
-  {
-    "city":"New York",
-    "state_region":" NY",
-    "country":" USA",
-    "lat":40.7127837,
-    "lng":-74.0059413
-  },
-
-  {
-    "city":"London",
-    "state_region":"",
-    "country":" UK",
-    "lat":51.5073509,
-    "lng":-0.1277583
-  },
-
-  {
-    "city":"Hong Kong",
-    "state_region": "",
-    "country":"Hong Kong",
-    "lat":22.396428,
-    "lng":114.109497
-  },
-
-  {
-    "city":"Rio de Janeiro",
-    "state_region":" State of Rio de Janeiro",
-    "country":" Brazil",
-    "lat":-22.9068467,
-    "lng":-43.1728965
-  },
-
-  {
-    "city":"Sydney NSW",
-    "state_region":"",
-    "country":" Australia",
-    "lat":-33.8688197,
-    "lng":151.2092955
-  },
-
-  {
-    "city":"Guangzhou",
-    "state_region":" Guangdong",
-    "country":" China",
-    "lat":23.12911,
-    "lng":113.264385
-  },
-
-  {
-    "city":"Singapore",
-    "state_region": "",
-    "country":"Singapore",
-    "lat":1.3553794,
-    "lng":103.8677444
-  },
-
-  {
-    "city":"Los Angeles",
-    "state_region":" CA",
-    "country":" USA",
-    "lat":34.0522342,
-    "lng":-118.2436849
-  },
-
-  {
-    "city":"Berlin",
-    "state_region":"",
-    "country":" Germany",
-    "lat":52.52000659999999,
-    "lng":13.404954
-  },
-
-  {
-    "city":"Madrid",
-    "state_region":"",
-    "country":" Spain",
-    "lat":40.4167754,
-    "lng":-3.7037902
-  },
-
-  {
-    "city":"Rome",
-    "state_region":"",
-    "country":" Italy",
-    "lat":41.9027835,
-    "lng":12.4963655
-  },
-
-  {
-    "city":"Taipei City",
-    "state_region":"",
-    "country":" Taiwan",
-    "lat":25.0329636,
-    "lng":121.5654268
-  },
-
-  {
-    "city":"Osaka",
-    "state_region":" Osaka Prefecture",
-    "country":" Japan",
-    "lat":34.6937378,
-    "lng":135.5021651
-  },
-
-  {
-    "city":"Toronto",
-    "state_region":" ON",
-    "country":" Canada",
-    "lat":43.653226,
-    "lng":-79.3831843
-  },
-
-  {
-    "city":"Paris",
-    "state_region":"",
-    "country":" France",
-    "lat":48.856614,
-    "lng":2.3522219
-  },
-
-  {
-    "city":"Houston",
-    "state_region":" TX",
-    "country":" USA",
-    "lat":29.7604267,
-    "lng":-95.3698028
-  },
-
-  {
-    "city":"Medellin",
-    "state_region":" Antioquia",
-    "country":" Colombia",
-    "lat":6.244203,
-    "lng":-75.5812119
-  },
-
-  {
-    "city":"Budapest",
-    "state_region":"",
-    "country":" Hungary",
-    "lat":47.497912,
-    "lng":19.040235
-  },
-
-  {
-    "city":"Dubai",
-    "state_region":"",
-    "country":"United Arab Emirates",
-    "lat":25.190248,
-    "lng":55.277778
-  },
-
-  {
-    "city":"Amsterdam",
-    "state_region":"",
-    "country":"Netherlands",
-    "lat":52.368631,
-    "lng":4.888371
-  }
-]
-
-
-def parseLatLng(n)
-  change = rand(0.1) * 0.023
-
-  randOp = rand(2)
-  if randOp == 0
-    return n + change
-  else
-    return n - change
-  end
-end
-
-20.times do
-  spot_locations.each_with_index do |spot_location|
-    spot_location[:lat] = parseLatLng(spot_location[:lat])
-    spot_location[:lng] = parseLatLng(spot_location[:lng])
-
-    city = spot_location[:city]
-
-    spot = Spot.new(spot_location)
-
-    title_adjs = [
-      'Beautiful',
-      'Spacious',
-      'Great',
-      'Wonderful',
-      'Best',
-      'Cheap',
-      'Luxurious'
-    ]
-    title_adj_idx = rand(title_adjs.count-1)
-    title_adj = title_adjs[title_adj_idx]
-
-    title_nouns = [
-      'Home',
-      'Room',
-      'Place',
-      'House',
-      'Apartment',
-      'Haven',
-      'Condo',
-      'Residence'
-    ]
-    title_noun_idx = rand(title_nouns.count-1)
-    title_noun = title_nouns[title_noun_idx]
-
-    room_types = [
-      'Entire House/Apt',
-      'Private Room',
-      'Shared Room'
-    ]
-    room_type_idx = rand(3)
-    room_type = room_types[room_type_idx]
-
-    post_code = rand(99999)
-    price_per_night = rand(1000) + 10
-
-    bed_count = rand(10) + 1
-    max_guests = rand(10) + 1
-
-    spot_pic_urls = [
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900936/spot17_qlwvzm.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900954/spot8_hecbj4.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot7_es4acz.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900935/spot18_moq6jj.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot12_rwluji.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900936/spot11_fytnsk.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot14_tdczmw.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot2_ukcmjb.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot19_x2exoh.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot13_av1dor.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot4_asyv66.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900940/spot15_ujvbug.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900941/spot9_tfltnp.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900942/spot1_mn0cen.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900956/chicago_sjvjc9.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900943/spot20_bqqzsx.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900961/spot10_mysua7.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900965/spot6_kmsylw.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900958/spot16_dfmr4l.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900956/spot3_ooxov4.jpg",
-      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900942/spot5_cwagxm.jpg"
-    ]
-    spot_pic_url_idx = rand(spot_pic_urls.count-1)
-    spot_pic_url = spot_pic_urls[spot_pic_url_idx]
-
-    gen_host_id = rand(67) + 2
-
-    spot.host_id = gen_host_id
-    spot.host_name = 'Host'
-    spot.title = "#{title_adj} #{title_noun} in #{city}"
-    spot.description = "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
-    spot.post_code = "#{post_code}"
-    spot.street_address = "12345 fake street address"
-    spot.price_per_night = price_per_night
-    spot.room_type = "#{room_type}"
-    spot.bed_count = bed_count
-    spot.max_guests = max_guests
-    spot.spot_pic_url = spot_pic_url
-
-    spot.save!
-  end
-end
+require_relative 'user_seeds'
+require_relative 'spot_seeds'
+require_relative 'review_seeds'
+
+
+# User.create({
+#   first_name: "Guest",
+#   last_name: "User",
+#   email: "guest@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482131647/couchbnb/person-solid_telh7f.png"
+# })
+#
+# User.create({
+#   first_name: "Jay",
+#   last_name: "Hwang",
+#   email: "jay_hwang@gmail.com",
+#   password: "password",
+#   profile_pic_url: "http://res.cloudinary.com/ddgtwtbre/image/upload/v1481587215/portfolio/linkedin_eiulzq.jpg"
+# })
+#
+# User.create({
+#   first_name: "Adam",
+#   last_name: "Foosaner",
+#   email: 'adam@gmail.com',
+#   password: 'password',
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Adam-Foosaner_nypijr.png"
+# })
+#
+# User.create({
+#   first_name: "Alex",
+#   last_name: "Bennet",
+#   email: "alex@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alex-Bennet_bpdgy6.png"
+# })
+#
+# User.create({
+#   first_name: "Alex",
+#   last_name: "Manning",
+#   email: "alex_manning@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alex-Manning_tzlryt.png"
+# })
+#
+# User.create({
+#   first_name: "Alex",
+#   last_name: "Doundakov",
+#   email: "alex_doundakov@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alexander-Doundakov_g045nm.png"
+# })
+#
+# User.create({
+#   first_name: "Amanda",
+#   last_name: "Fielding",
+#   email: "amanda@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128657/couchbnb/Amanda-Fielding_fnmudx.png"
+# })
+#
+# User.create({
+#   first_name: "Andrew",
+#   last_name: "Tae",
+#   email: "andrew_tae@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Andrew-Tae_ht1kcz.png"
+# })
+#
+# User.create({
+#   first_name: "Anudeep",
+#   last_name: "Kasturi",
+#   email: "anudeep_kasturi@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128664/couchbnb/Anudeep-Kasturi_odxgby.png"
+# })
+#
+# User.create({
+#   first_name: "Benjamin",
+#   last_name: "Dippelsman",
+#   email: "ben_dip@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128658/couchbnb/Benjamin-Dippelsman_xl0tbl.png"
+# })
+#
+# User.create({
+#   first_name: "Brian",
+#   last_name: "Chen",
+#   email: "brian_chen@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/Brian-Chen_mwcxww.png"
+# })
+#
+# User.create({
+#   first_name: "Brian",
+#   last_name: "Lee",
+#   email: "brian_lee@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128666/couchbnb/Brian-Lee_l4b2h5.png"
+# })
+#
+# User.create({
+#   first_name: "Brian",
+#   last_name: "Tsai",
+#   email: "brian_tsai@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Brian-Tsai_n5qaky.png"
+# })
+#
+# User.create({
+#   first_name: "Caleb",
+#   last_name: "Ontiveros",
+#   email: "caleb_onti@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Caleb-Ontiveros_ihuvpj.png"
+# })
+#
+# User.create({
+#   first_name: "Christopher",
+#   last_name: "Button",
+#   email: "christopher_button@gmail.com",
+#   password: 'password',
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/Christopher-Button_o3lysh.png"
+# })
+#
+# User.create({
+#   first_name: "Cindy",
+#   last_name: "Wang",
+#   email: "cindy_wang@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/Cindy-Wang_woewj1.png"
+# })
+#
+# User.create({
+#   first_name: "Connor",
+#   last_name: "Murphy",
+#   email: "connor_muphy@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128668/couchbnb/Connor-Murphy_kmoitq.png"
+# })
+#
+# User.create({
+#   first_name: "Daniel",
+#   last_name: "Sulfaro",
+#   email: "daniel_sulfaro@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128667/couchbnb/Daniel-Sulfaro_lnsv0o.png"
+# })
+#
+# User.create({
+#   first_name: "David",
+#   last_name: "Kim",
+#   email: "david_kim@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/David-Kim_nqb3b7.png"
+# })
+#
+# User.create({
+#   first_name: "David",
+#   last_name: "McAfee",
+#   email: 'david_mcafee@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/David-McAfee_v2okis.png"
+# })
+#
+# User.create({
+#   first_name: "Debra",
+#   last_name: "Fong",
+#   email: "debra_fong@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128668/couchbnb/Debra-Fong_junc34.png"
+# })
+#
+# User.create({
+#   first_name: "Deep",
+#   last_name: "Tailor",
+#   email: "deep_tailor@gmail.com",
+#   password: "password",
+#   profile_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/Deep-Tailor_fcid09.png'
+# })
+#
+# User.create({
+#   first_name: "Duc",
+#   last_name: "Nguyen",
+#   email: "duc_nguyen@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128670/couchbnb/Duc-Nguyen_tqn8ll.png"
+# })
+#
+# User.create({
+#   first_name: "Eric",
+#   last_name: "Kwok",
+#   email: "eric_kwok@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/Eric-Kwok_tijkgu.png"
+# })
+#
+# User.create({
+#   first_name: "Eric",
+#   last_name: "Oliveira",
+#   email: 'eric_oliveira@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Eric-Oliveira_wpe6od.png"
+# })
+#
+# User.create({
+#   first_name: "Ethan",
+#   last_name: "Wan",
+#   email: "ethan_wan@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Ethan-Wan_ekbgwv.png"
+# })
+#
+# User.create({
+#   first_name: "Evan",
+#   last_name: "Waits",
+#   email: "evan_waits@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Ethan-Wan_ekbgwv.png"
+# })
+#
+# User.create({
+#   first_name: "Fan",
+#   last_name: "Song",
+#   email: "fan_song@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Fan-Song_ul7irn.png"
+# })
+#
+# User.create({
+#   first_name: "Frank",
+#   last_name: "Bi",
+#   email: 'frank_bi@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Frank-Bi_yczosf.png"
+# })
+#
+# User.create({
+#   first_name: "Gabriel",
+#   last_name: "Lewis",
+#   email: "gabriel_lewis@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128673/couchbnb/Gabriel-Lewis_srgc8n.png"
+# })
+#
+# User.create({
+#   first_name: "Humza",
+#   last_name: "Baig",
+#   email: 'humza_baig@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Hamza-Baig_jnvfo1.png"
+# })
+#
+# User.create({
+#   first_name: "Henry",
+#   last_name: "Hsu",
+#   email: "henry_hsu@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Hamza-Baig_jnvfo1.png"
+# })
+#
+# User.create({
+#   first_name: "James",
+#   last_name: "Conklin",
+#   email: "james_conklin",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128678/couchbnb/James-Conklin_p0dvzr.png"
+# })
+#
+# User.create({
+#   first_name: "Jane",
+#   last_name: "Hauf",
+#   email: "jane_hauf@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Jane-Hauf_chq1eb.png"
+# })
+#
+# User.create({
+#   first_name: "Janine",
+#   last_name: "Love",
+#   email: 'janine_love@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128676/couchbnb/Janine-Love_fjhlft.png"
+# })
+#
+# User.create({
+#   first_name: 'JC',
+#   last_name: 'Balcita',
+#   email: "jc_balcita@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128679/couchbnb/JC-Balcita_uhjc9l.png"
+# })
+#
+# User.create({
+#   first_name: "John",
+#   last_name: "Lekili",
+#   email: 'john_lekili@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128678/couchbnb/John-Lekili_hjymxe.png"
+# })
+#
+# User.create({
+#   first_name: "Joseph",
+#   last_name: "McBride",
+#   email: 'joseph_mcbride@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128681/couchbnb/Joseph-McBride_rt9bro.png"
+# })
+#
+# User.create({
+#   first_name: "Justice",
+#   last_name: "Yen",
+#   email: "justice_yen@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128679/couchbnb/Justice-Yen_vaqjbg.png"
+# })
+#
+# User.create({
+#   first_name: "Lindsey",
+#   last_name: "Whitley",
+#   email: 'lindsey_whitley@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Lindsey-Whitley_pwedp2.png"
+# })
+#
+# User.create({
+#   first_name: "Louis",
+#   last_name: "Cruz",
+#   email: "louis_cruz@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Louis-Cruz_bwx4za.png"
+# })
+#
+# User.create({
+#   first_name: "Luke",
+#   last_name: "Wassink",
+#   email: "luke_wassink@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Luke-Wassink_dvzgti.png"
+# })
+#
+# User.create({
+#   first_name: "Matt",
+#   last_name: "Velker",
+#   email: "matt_velker@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128680/couchbnb/Matt-Velker_snyprp.png"
+# })
+#
+# User.create({
+#   first_name: "Melissa",
+#   last_name: "Lau",
+#   email: "melissa_lau@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128683/couchbnb/Melissa-Lau_xrxrbh.png"
+# })
+#
+# User.create({
+#   first_name: "Micah",
+#   last_name: "Sapitsky",
+#   email: "micach_sapitsky@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128684/couchbnb/Micah-Sapitsky_yt84jv.png"
+# })
+#
+# User.create({
+#   first_name: "Michael",
+#   last_name: "Hirabayashi",
+#   email: "michael_hirabayashi@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Michael-Hirabayashi_jd05gl.png"
+# })
+#
+# User.create({
+#   first_name: "Michael",
+#   last_name: "Loschiavo",
+#   email: 'michael_loschiavo@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128688/couchbnb/Michael-Loschiavo_kzxqhl.png"
+# })
+#
+# User.create({
+#   first_name: "Michael",
+#   last_name: "Pak",
+#   email: "michael_pak@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Michael-Pak_dqyesg.png"
+# })
+#
+# User.create({
+#   first_name: "Michaelangelo",
+#   last_name: "Loreti",
+#   email: "miki_loreti@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Michaelangelo-Loreti_td9ell.png"
+# })
+#
+# User.create({
+#   first_name: 'Mike',
+#   last_name: "Barille",
+#   email: 'mike_barille@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128687/couchbnb/Mike-Barille_igrjta.png"
+# })
+#
+# User.create({
+#   first_name: "Miki",
+#   last_name: "Sparks",
+#   email: "miki_sparks@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128687/couchbnb/Miki-Sparks_yttcgo.png"
+# })
+#
+# User.create({
+#   first_name: "Nicholas",
+#   last_name: "Skoularikis",
+#   email: 'nicho@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Nicholas-Skoularikis_p71di2.png"
+# })
+#
+# User.create({
+#   first_name: "Paul",
+#   last_name: "Cathey",
+#   email: "paul_cathey@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128688/couchbnb/Paul-Cathey_uqbi6h.png"
+# })
+#
+# User.create({
+#   first_name: "Peik",
+#   last_name: "Sia",
+#   email: "peik_sia@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Peik-Sia_bqpfzr.png"
+# })
+#
+# User.create({
+#   first_name: "Pysith",
+#   last_name: "Vanuptikul",
+#   email: 'pi_van@gmail.com',
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Pysith_Vanuptikul_eh6roj.png"
+# })
+#
+# User.create({
+#   first_name: "Samuel",
+#   last_name: "Mak",
+#   email: "sam_mak@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128692/couchbnb/Samuel-Mak_e9huh3.png"
+# })
+#
+# User.create({
+#   first_name: "Sebastian",
+#   last_name: "Cruz",
+#   email: "seb_cruz@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128691/couchbnb/Sebastian-Cruz_lm5qde.png"
+# })
+#
+# User.create({
+#   first_name: "Shaurya",
+#   last_name: "Sinha",
+#   email: "shaurya_sinha@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128690/couchbnb/Shaurya-Sinha_fbwey8.png"
+# })
+#
+# User.create({
+#   first_name: "Sonik",
+#   last_name: "Jang",
+#   email: "sonik_jang@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Sonik-Jhang_ajrt0r.png"
+# })
+#
+# User.create({
+#   first_name: "Taylor",
+#   last_name: "Chestnut",
+#   email: "taylor_chestnut",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128691/couchbnb/Taylor-Chestnut_wkwtrx.png"
+# })
+#
+# User.create({
+#   first_name: "Todd",
+#   last_name: "Nestor",
+#   email: "todd_nestor@gmail.com",
+#   password: "password",
+#   profile_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128693/Todd-Nestor_ehdiai.png'
+# })
+#
+# User.create({
+#   first_name: "Tonia",
+#   last_name: "Hsia",
+#   email: "tonia_hsia@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Tonia-Hsia_dxh834.png"
+# })
+#
+# User.create({
+#   first_name: "Tony",
+#   last_name: "Xie",
+#   email: "tony_xie@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Tony-Xie_prap45.png"
+# })
+#
+# User.create({
+#   first_name: "Travis",
+#   last_name: "Ludlum",
+#   email: "travis_ludlum@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128693/couchbnb/Travis-Ludlum_dnjeqg.png"
+# })
+#
+# User.create({
+#   first_name: "Trevor",
+#   last_name: "Scandalios",
+#   email: "trevor_scan@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Trevor-Scandalios_lwkfma.png"
+# })
+#
+# User.create({
+#   first_name: "Tyler",
+#   last_name: "Fields",
+#   email: "tyler_fields@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Trevor-Scandalios_lwkfma.png"
+# })
+#
+# User.create({
+#   first_name: "Victor",
+#   last_name: "Aw",
+#   email: "victor_aw@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Victor-Aw_vesrpm.png"
+# })
+#
+# User.create({
+#   first_name: "Vince",
+#   last_name: "Sangke",
+#   email: "vince_sangke@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128695/couchbnb/Vince-Sangkhae_gotqet.png"
+# })
+#
+# User.create({
+#   first_name: "Winston",
+#   last_name: "Zhao",
+#   email: "winston_zhao@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Winston-Zhao_a5pqvw.png"
+# })
+#
+# User.create({
+#   first_name: "Yue",
+#   last_name: "Gu",
+#   email: "yue_gu@gmail.com",
+#   password: "password",
+#   profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Yue-Gu_nxn82n.png"
+# })
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Host",
+#   title: "San Francisco Apt Downtown",
+#   description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+#   country: "United States",
+#   state_region: "California",
+#   city: "San Francisco",
+#   post_code: '94103',
+#   street_address: '12345 California St',
+#   price_per_night: 100,
+#   room_type: "Entire House/Apt",
+#   bed_count: 4,
+#   max_guests: 4,
+#   spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic3_hdfvpl.jpg',
+#   lat: 37.794144,
+#   lng: -122.408957
+# })
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Jay",
+#   title: "Los Angeles Apt Downtown",
+#   description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+#   country: "United States",
+#   state_region: "California",
+#   city: "Los Angeles",
+#   post_code: '91002',
+#   street_address: '12345 Wilshire Blvd',
+#   price_per_night: 100,
+#   room_type: "Entire House/Apt",
+#   bed_count: 4,
+#   max_guests: 4,
+#   spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic1_ojh6w2.jpg',
+#   lat: 34.052256,
+#   lng: -118.254381
+# })
+#
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Jay",
+#   title: "Private Studio in London",
+#   description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+#   country: "United Kingdom",
+#   state_region: "",
+#   city: "London",
+#   post_code: 'AG42NB',
+#   street_address: '12345 California St',
+#   price_per_night: 100,
+#   room_type: "Private Room",
+#   bed_count: 1,
+#   max_guests: 2,
+#   spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic4_r0oafw.jpg',
+#   lat: 51.520915,
+#   lng: -0.149832,
+# })
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Jay",
+#   title: "Quaint room in Los Angeles",
+#   description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+#   country: "United States",
+#   state_region: "California",
+#   city: "Los Angeles",
+#   post_code: '91002',
+#   street_address: '12345 Sunset Blvd',
+#   price_per_night: 50,
+#   room_type: "Shared Room",
+#   bed_count: 2,
+#   max_guests: 4,
+#   spot_pic_url: 'http://res.cloudinary.com/ddgtwtbre/image/upload/v1478492826/spot1pic2_ucamo9.jpg',
+#   lat: 34.090932,
+#   lng: -118.406058
+# })
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Jay",
+#   title: "Disneyland Paris perfect for 5!",
+#   description: "This apartment is ideally located at the foot of the famous Valley Shopping and just close to the Disney Parks!! Finish a day full of emotions for a moment of relaxation in the pool of the residence.",
+#   country: "France",
+#   state_region: "Île-de-France",
+#   city: "Serris",
+#   post_code: "77700",
+#   street_address: "12345 Paris Ave",
+#   price_per_night: 75,
+#   room_type: "Entire House/Apt",
+#   bed_count: 2,
+#   max_guests: 5,
+#   spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/c_scale,h_604/v1478628133/paris-eiffel-tower_hxpxjx.jpg',
+#   lat: 48.843499,
+#   lng: 2.785134
+# })
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Jay",
+#   title: "Luxury Flat",
+#   description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+#   country: "Netherlands",
+#   state_region: "Noord-Holland",
+#   city: "Amsterdam",
+#   post_code: "12010",
+#   street_address: "12345 Amsterdam Street",
+#   price_per_night: 147,
+#   room_type: "Entire House/Apt",
+#   bed_count: 6,
+#   max_guests: 8,
+#   spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478628600/amsterdam-luxury-flat_jkgxej.webp',
+#   lat: 52.374172,
+#   lng: 4.861534
+# })
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Jay",
+#   title: "Entertainer's Paradise, Heart of Umhlanga",
+#   description: "Our contemporary home boasts open plan living at its finest. This fully furnished home is situated in a quiet and safe complex in Umhlanga, Durban. The house consists of four en-suite bedrooms, a fully equipped kitchen, 12 seating dining area & 2 TV lounges. There is a large infinity pool with heated Jacuzzi located off a sunset facing pool deck looking into natural bush.
+#
+#   A 2 minute drive to the beach, our home is the perfect escape.",
+#   country: "South Africa",
+#   state_region: 'KZN',
+#   city: "Durban",
+#   post_code: '12345',
+#   street_address: "12345 Durban Poison Ave",
+#   price_per_night: 801,
+#   room_type: "Entire House/Apt",
+#   bed_count: 5,
+#   max_guests: 8,
+#   spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478629001/durban-south-africa_yp7win.webp',
+#   lat: -29.856882,
+#   lng: 31.036159
+# })
+#
+# Spot.create({
+#   host_id: 2,
+#   host_name: "Jay",
+#   title: "Spacious Home in heart of Seoul",
+#   description: "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+#   country: "South Korea",
+#   state_region: "",
+#   city: "Seoul",
+#   post_code: '09876',
+#   street_address: "1234 Seoul Street",
+#   price_per_night: 120,
+#   room_type: "Entire House/Apt",
+#   bed_count: 3,
+#   max_guests: 6,
+#   spot_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1478629386/seoul-south-korea_a1lsjg.jpg',
+#   lat: 37.513837,
+#   lng: 127.029368
+# })
+#
+# spot_locations = [
+#   {
+#     "city":"San Francisco",
+#     "state_region":" California",
+#     "country":" USA",
+#     "lat":37.7609082,
+#     "lng":-122.4350043
+#   },
+#
+#   {
+#     "city":"Los Angeles",
+#     "state_region":" CA",
+#     "country":" USA",
+#     "lat":34.0522342,
+#     "lng":-118.2436849
+#   },
+#
+#   {
+#     "city":"Hollywood",
+#     "state_region":" Los Angeles",
+#     "country":" USA",
+#     "lat":34.0928092,
+#     "lng":-118.3286614
+#   },
+#
+#   {
+#     "city":"Phoenix",
+#     "state_region":" AZ",
+#     "country":" USA",
+#     "lat":33.4483771,
+#     "lng":-112.0740373
+#   },
+#
+#   {
+#     "city":"Tucson",
+#     "state_region":" AZ",
+#     "country":" USA",
+#     "lat":32.2217429,
+#     "lng":-110.926479
+#   },
+#
+#   {
+#     "city":"El Paso",
+#     "state_region":" TX",
+#     "country":" USA",
+#     "lat":31.7618778,
+#     "lng":-106.4850217
+#   },
+#
+#   {
+#     "city":"Brooklyn",
+#     "state_region":" NY",
+#     "country":" USA",
+#     "lat":40.6781784,
+#     "lng":-73.9441579
+#   },
+#
+#   {
+#     "city":"Bronx County",
+#     "state_region":" NY",
+#     "country":" USA",
+#     "lat":40.8370495,
+#     "lng":-73.86542949999999
+#   },
+#
+#   {
+#     "city":"Queens County",
+#     "state_region":" NY",
+#     "country":" USA",
+#     "lat":40.7282239,
+#     "lng":-73.7948516
+#   },
+#
+#   {
+#     "city":"Manhattan",
+#     "state_region":" New York",
+#     "country":" USA",
+#     "lat":40.7830603,
+#     "lng":-73.9712488
+#   },
+#
+#   {
+#     "city":"Chicago",
+#     "state_region":" IL",
+#     "country":" USA",
+#     "lat":41.8781136,
+#     "lng":-87.6297982
+#   },
+#
+#   {
+#     "city":"Las Vegas",
+#     "state_region":" NV",
+#     "country":" USA",
+#     "lat":36.1699412,
+#     "lng":-115.1398296
+#   },
+#
+#   {
+#     "city":"Seattle",
+#     "state_region":" WA",
+#     "country":" USA",
+#     "lat":47.6062095,
+#     "lng":-122.3320708
+#   },
+#
+#   {
+#     "city":"Washington",
+#     "state_region":" DC",
+#     "country":" USA",
+#     "lat":38.9071923,
+#     "lng":-77.0368707
+#   },
+#
+#   {
+#     "city":"Miami",
+#     "state_region":" FL",
+#     "country":" USA",
+#     "lat":25.7616798,
+#     "lng":-80.1917902
+#   },
+#
+#   {
+#     "city":"Boston",
+#     "state_region":" MA",
+#     "country":" USA",
+#     "lat":42.3600825,
+#     "lng":-71.0588801
+#   },
+#
+#   {
+#     "city":"Portland",
+#     "state_region":" OR",
+#     "country":" USA",
+#     "lat":45.5230622,
+#     "lng":-122.6764816
+#   },
+#
+#   {
+#     "city":"Nashville",
+#     "state_region":" TN",
+#     "country":" USA",
+#     "lat":36.1626638,
+#     "lng":-86.7816016
+#   },
+#
+#   {
+#     "city":"Dallas",
+#     "state_region":" TX",
+#     "country":" USA",
+#     "lat":32.7766642,
+#     "lng":-96.79698789999999
+#   },
+#
+#   {
+#     "city":"Austin",
+#     "state_region":" TX",
+#     "country":" USA",
+#     "lat":30.267153,
+#     "lng":-97.7430608
+#   },
+#
+#   {
+#     "city":"Minneapolis",
+#     "state_region":" MN",
+#     "country":" USA",
+#     "lat":44.977753,
+#     "lng":-93.2650108
+#   },
+#
+#   {
+#     "city":"Seoul",
+#     "state_region":"",
+#     "country":" South Korea",
+#     "lat":37.566535,
+#     "lng":126.9779692
+#   },
+#
+#   {
+#     "city":"São Paulo",
+#     "state_region":" State of São Paulo",
+#     "country":" Brazil",
+#     "lat":-23.5505199,
+#     "lng":-46.63330939999999
+#   },
+#
+#   {
+#     "city":"Mumbai",
+#     "state_region":" Maharashtra",
+#     "country":" India",
+#     "lat":19.0759837,
+#     "lng":72.8776559
+#   },
+#
+#   {
+#     "city":"Moscow",
+#     "state_region":"",
+#     "country":" Russia",
+#     "lat":55.755826,
+#     "lng":37.6173
+#   },
+#
+#   {
+#     "city":"Shanghai",
+#     "state_region":"",
+#     "country":" China",
+#     "lat":31.230416,
+#     "lng":121.473701
+#   },
+#
+#   {
+#     "city":"Tokyo",
+#     "state_region":"",
+#     "country":" Japan",
+#     "lat":35.7090259,
+#     "lng":139.7319925
+#   },
+#
+#   {
+#     "city":"New York",
+#     "state_region":" NY",
+#     "country":" USA",
+#     "lat":40.7127837,
+#     "lng":-74.0059413
+#   },
+#
+#   {
+#     "city":"London",
+#     "state_region":"",
+#     "country":" UK",
+#     "lat":51.5073509,
+#     "lng":-0.1277583
+#   },
+#
+#   {
+#     "city":"Hong Kong",
+#     "state_region": "",
+#     "country":"Hong Kong",
+#     "lat":22.396428,
+#     "lng":114.109497
+#   },
+#
+#   {
+#     "city":"Rio de Janeiro",
+#     "state_region":" State of Rio de Janeiro",
+#     "country":" Brazil",
+#     "lat":-22.9068467,
+#     "lng":-43.1728965
+#   },
+#
+#   {
+#     "city":"Sydney NSW",
+#     "state_region":"",
+#     "country":" Australia",
+#     "lat":-33.8688197,
+#     "lng":151.2092955
+#   },
+#
+#   {
+#     "city":"Guangzhou",
+#     "state_region":" Guangdong",
+#     "country":" China",
+#     "lat":23.12911,
+#     "lng":113.264385
+#   },
+#
+#   {
+#     "city":"Singapore",
+#     "state_region": "",
+#     "country":"Singapore",
+#     "lat":1.3553794,
+#     "lng":103.8677444
+#   },
+#
+#   {
+#     "city":"Los Angeles",
+#     "state_region":" CA",
+#     "country":" USA",
+#     "lat":34.0522342,
+#     "lng":-118.2436849
+#   },
+#
+#   {
+#     "city":"Berlin",
+#     "state_region":"",
+#     "country":" Germany",
+#     "lat":52.52000659999999,
+#     "lng":13.404954
+#   },
+#
+#   {
+#     "city":"Madrid",
+#     "state_region":"",
+#     "country":" Spain",
+#     "lat":40.4167754,
+#     "lng":-3.7037902
+#   },
+#
+#   {
+#     "city":"Rome",
+#     "state_region":"",
+#     "country":" Italy",
+#     "lat":41.9027835,
+#     "lng":12.4963655
+#   },
+#
+#   {
+#     "city":"Taipei City",
+#     "state_region":"",
+#     "country":" Taiwan",
+#     "lat":25.0329636,
+#     "lng":121.5654268
+#   },
+#
+#   {
+#     "city":"Osaka",
+#     "state_region":" Osaka Prefecture",
+#     "country":" Japan",
+#     "lat":34.6937378,
+#     "lng":135.5021651
+#   },
+#
+#   {
+#     "city":"Toronto",
+#     "state_region":" ON",
+#     "country":" Canada",
+#     "lat":43.653226,
+#     "lng":-79.3831843
+#   },
+#
+#   {
+#     "city":"Paris",
+#     "state_region":"",
+#     "country":" France",
+#     "lat":48.856614,
+#     "lng":2.3522219
+#   },
+#
+#   {
+#     "city":"Houston",
+#     "state_region":" TX",
+#     "country":" USA",
+#     "lat":29.7604267,
+#     "lng":-95.3698028
+#   },
+#
+#   {
+#     "city":"Medellin",
+#     "state_region":" Antioquia",
+#     "country":" Colombia",
+#     "lat":6.244203,
+#     "lng":-75.5812119
+#   },
+#
+#   {
+#     "city":"Budapest",
+#     "state_region":"",
+#     "country":" Hungary",
+#     "lat":47.497912,
+#     "lng":19.040235
+#   },
+#
+#   {
+#     "city":"Dubai",
+#     "state_region":"",
+#     "country":"United Arab Emirates",
+#     "lat":25.190248,
+#     "lng":55.277778
+#   },
+#
+#   {
+#     "city":"Amsterdam",
+#     "state_region":"",
+#     "country":"Netherlands",
+#     "lat":52.368631,
+#     "lng":4.888371
+#   }
+# ]
+#
+#
+# def parseLatLng(n)
+#   change = rand(0.1) * 0.023
+#
+#   randOp = rand(2)
+#   if randOp == 0
+#     return n + change
+#   else
+#     return n - change
+#   end
+# end
+#
+# 20.times do
+#   spot_locations.each_with_index do |spot_location|
+#     spot_location[:lat] = parseLatLng(spot_location[:lat])
+#     spot_location[:lng] = parseLatLng(spot_location[:lng])
+#
+#     city = spot_location[:city]
+#
+#     spot = Spot.new(spot_location)
+#
+#     title_adjs = [
+#       'Beautiful',
+#       'Spacious',
+#       'Great',
+#       'Wonderful',
+#       'Best',
+#       'Cheap',
+#       'Luxurious'
+#     ]
+#     title_adj_idx = rand(title_adjs.count-1)
+#     title_adj = title_adjs[title_adj_idx]
+#
+#     title_nouns = [
+#       'Home',
+#       'Room',
+#       'Place',
+#       'House',
+#       'Apartment',
+#       'Haven',
+#       'Condo',
+#       'Residence'
+#     ]
+#     title_noun_idx = rand(title_nouns.count-1)
+#     title_noun = title_nouns[title_noun_idx]
+#
+#     room_types = [
+#       'Entire House/Apt',
+#       'Private Room',
+#       'Shared Room'
+#     ]
+#     room_type_idx = rand(3)
+#     room_type = room_types[room_type_idx]
+#
+#     post_code = rand(99999)
+#     price_per_night = rand(1000) + 10
+#
+#     bed_count = rand(10) + 1
+#     max_guests = rand(10) + 1
+#
+#     spot_pic_urls = [
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900936/spot17_qlwvzm.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900954/spot8_hecbj4.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot7_es4acz.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900935/spot18_moq6jj.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot12_rwluji.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900936/spot11_fytnsk.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot14_tdczmw.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot2_ukcmjb.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot19_x2exoh.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot13_av1dor.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot4_asyv66.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900940/spot15_ujvbug.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900941/spot9_tfltnp.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900942/spot1_mn0cen.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900956/chicago_sjvjc9.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900943/spot20_bqqzsx.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900961/spot10_mysua7.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900965/spot6_kmsylw.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900958/spot16_dfmr4l.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900956/spot3_ooxov4.jpg",
+#       "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900942/spot5_cwagxm.jpg"
+#     ]
+#     spot_pic_url_idx = rand(spot_pic_urls.count-1)
+#     spot_pic_url = spot_pic_urls[spot_pic_url_idx]
+#
+#     gen_host_id = rand(67) + 2
+#
+#     spot.host_id = gen_host_id
+#     spot.host_name = 'Host'
+#     spot.title = "#{title_adj} #{title_noun} in #{city}"
+#     spot.description = "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
+#     spot.post_code = "#{post_code}"
+#     spot.street_address = "12345 fake street address"
+#     spot.price_per_night = price_per_night
+#     spot.room_type = "#{room_type}"
+#     spot.bed_count = bed_count
+#     spot.max_guests = max_guests
+#     spot.spot_pic_url = spot_pic_url
+#
+#     spot.save!
+#   end
+# end
