+++ db/spot_seeds.rb
@@ -1,0 +1,481 @@
+spot_locations = [
+  {
+    "city":"San Francisco",
+    "state_region":" California",
+    "country":" USA",
+    "lat":37.7609082,
+    "lng":-122.4350043
+  },
+
+  {
+    "city":"Los Angeles",
+    "state_region":" CA",
+    "country":" USA",
+    "lat":34.0522342,
+    "lng":-118.2436849
+  },
+
+  {
+    "city":"Hollywood",
+    "state_region":" Los Angeles",
+    "country":" USA",
+    "lat":34.0928092,
+    "lng":-118.3286614
+  },
+
+  {
+    "city":"Phoenix",
+    "state_region":" AZ",
+    "country":" USA",
+    "lat":33.4483771,
+    "lng":-112.0740373
+  },
+
+  {
+    "city":"Tucson",
+    "state_region":" AZ",
+    "country":" USA",
+    "lat":32.2217429,
+    "lng":-110.926479
+  },
+
+  {
+    "city":"El Paso",
+    "state_region":" TX",
+    "country":" USA",
+    "lat":31.7618778,
+    "lng":-106.4850217
+  },
+
+  {
+    "city":"Brooklyn",
+    "state_region":" NY",
+    "country":" USA",
+    "lat":40.6781784,
+    "lng":-73.9441579
+  },
+
+  {
+    "city":"Bronx County",
+    "state_region":" NY",
+    "country":" USA",
+    "lat":40.8370495,
+    "lng":-73.86542949999999
+  },
+
+  {
+    "city":"Queens County",
+    "state_region":" NY",
+    "country":" USA",
+    "lat":40.7282239,
+    "lng":-73.7948516
+  },
+
+  {
+    "city":"Manhattan",
+    "state_region":" New York",
+    "country":" USA",
+    "lat":40.7830603,
+    "lng":-73.9712488
+  },
+
+  {
+    "city":"Chicago",
+    "state_region":" IL",
+    "country":" USA",
+    "lat":41.8781136,
+    "lng":-87.6297982
+  },
+
+  {
+    "city":"Las Vegas",
+    "state_region":" NV",
+    "country":" USA",
+    "lat":36.1699412,
+    "lng":-115.1398296
+  },
+
+  {
+    "city":"Seattle",
+    "state_region":" WA",
+    "country":" USA",
+    "lat":47.6062095,
+    "lng":-122.3320708
+  },
+
+  {
+    "city":"Washington",
+    "state_region":" DC",
+    "country":" USA",
+    "lat":38.9071923,
+    "lng":-77.0368707
+  },
+
+  {
+    "city":"Miami",
+    "state_region":" FL",
+    "country":" USA",
+    "lat":25.7616798,
+    "lng":-80.1917902
+  },
+
+  {
+    "city":"Boston",
+    "state_region":" MA",
+    "country":" USA",
+    "lat":42.3600825,
+    "lng":-71.0588801
+  },
+
+  {
+    "city":"Portland",
+    "state_region":" OR",
+    "country":" USA",
+    "lat":45.5230622,
+    "lng":-122.6764816
+  },
+
+  {
+    "city":"Nashville",
+    "state_region":" TN",
+    "country":" USA",
+    "lat":36.1626638,
+    "lng":-86.7816016
+  },
+
+  {
+    "city":"Dallas",
+    "state_region":" TX",
+    "country":" USA",
+    "lat":32.7766642,
+    "lng":-96.79698789999999
+  },
+
+  {
+    "city":"Austin",
+    "state_region":" TX",
+    "country":" USA",
+    "lat":30.267153,
+    "lng":-97.7430608
+  },
+
+  {
+    "city":"Minneapolis",
+    "state_region":" MN",
+    "country":" USA",
+    "lat":44.977753,
+    "lng":-93.2650108
+  },
+
+  {
+    "city":"Seoul",
+    "state_region":"",
+    "country":" South Korea",
+    "lat":37.566535,
+    "lng":126.9779692
+  },
+
+  {
+    "city":"São Paulo",
+    "state_region":" State of São Paulo",
+    "country":" Brazil",
+    "lat":-23.5505199,
+    "lng":-46.63330939999999
+  },
+
+  {
+    "city":"Mumbai",
+    "state_region":" Maharashtra",
+    "country":" India",
+    "lat":19.0759837,
+    "lng":72.8776559
+  },
+
+  {
+    "city":"Moscow",
+    "state_region":"",
+    "country":" Russia",
+    "lat":55.755826,
+    "lng":37.6173
+  },
+
+  {
+    "city":"Shanghai",
+    "state_region":"",
+    "country":" China",
+    "lat":31.230416,
+    "lng":121.473701
+  },
+
+  {
+    "city":"Tokyo",
+    "state_region":"",
+    "country":" Japan",
+    "lat":35.7090259,
+    "lng":139.7319925
+  },
+
+  {
+    "city":"New York",
+    "state_region":" NY",
+    "country":" USA",
+    "lat":40.7127837,
+    "lng":-74.0059413
+  },
+
+  {
+    "city":"London",
+    "state_region":"",
+    "country":" UK",
+    "lat":51.5073509,
+    "lng":-0.1277583
+  },
+
+  {
+    "city":"Hong Kong",
+    "state_region": "",
+    "country":"Hong Kong",
+    "lat":22.396428,
+    "lng":114.109497
+  },
+
+  {
+    "city":"Rio de Janeiro",
+    "state_region":" State of Rio de Janeiro",
+    "country":" Brazil",
+    "lat":-22.9068467,
+    "lng":-43.1728965
+  },
+
+  {
+    "city":"Sydney NSW",
+    "state_region":"",
+    "country":" Australia",
+    "lat":-33.8688197,
+    "lng":151.2092955
+  },
+
+  {
+    "city":"Guangzhou",
+    "state_region":" Guangdong",
+    "country":" China",
+    "lat":23.12911,
+    "lng":113.264385
+  },
+
+  {
+    "city":"Singapore",
+    "state_region": "",
+    "country":"Singapore",
+    "lat":1.3553794,
+    "lng":103.8677444
+  },
+
+  {
+    "city":"Los Angeles",
+    "state_region":" CA",
+    "country":" USA",
+    "lat":34.0522342,
+    "lng":-118.2436849
+  },
+
+  {
+    "city":"Berlin",
+    "state_region":"",
+    "country":" Germany",
+    "lat":52.52000659999999,
+    "lng":13.404954
+  },
+
+  {
+    "city":"Madrid",
+    "state_region":"",
+    "country":" Spain",
+    "lat":40.4167754,
+    "lng":-3.7037902
+  },
+
+  {
+    "city":"Rome",
+    "state_region":"",
+    "country":" Italy",
+    "lat":41.9027835,
+    "lng":12.4963655
+  },
+
+  {
+    "city":"Taipei City",
+    "state_region":"",
+    "country":" Taiwan",
+    "lat":25.0329636,
+    "lng":121.5654268
+  },
+
+  {
+    "city":"Osaka",
+    "state_region":" Osaka Prefecture",
+    "country":" Japan",
+    "lat":34.6937378,
+    "lng":135.5021651
+  },
+
+  {
+    "city":"Toronto",
+    "state_region":" ON",
+    "country":" Canada",
+    "lat":43.653226,
+    "lng":-79.3831843
+  },
+
+  {
+    "city":"Paris",
+    "state_region":"",
+    "country":" France",
+    "lat":48.856614,
+    "lng":2.3522219
+  },
+
+  {
+    "city":"Houston",
+    "state_region":" TX",
+    "country":" USA",
+    "lat":29.7604267,
+    "lng":-95.3698028
+  },
+
+  {
+    "city":"Medellin",
+    "state_region":" Antioquia",
+    "country":" Colombia",
+    "lat":6.244203,
+    "lng":-75.5812119
+  },
+
+  {
+    "city":"Budapest",
+    "state_region":"",
+    "country":" Hungary",
+    "lat":47.497912,
+    "lng":19.040235
+  },
+
+  {
+    "city":"Dubai",
+    "state_region":"",
+    "country":"United Arab Emirates",
+    "lat":25.190248,
+    "lng":55.277778
+  },
+
+  {
+    "city":"Amsterdam",
+    "state_region":"",
+    "country":"Netherlands",
+    "lat":52.368631,
+    "lng":4.888371
+  }
+]
+
+
+def parseLatLng(n)
+  change = rand(0.1) * 0.023
+
+  randOp = rand(2)
+  if randOp == 0
+    return n + change
+  else
+    return n - change
+  end
+end
+
+20.times do
+  spot_locations.each_with_index do |spot_location|
+    spot_location[:lat] = parseLatLng(spot_location[:lat])
+    spot_location[:lng] = parseLatLng(spot_location[:lng])
+
+    city = spot_location[:city]
+
+    spot = Spot.new(spot_location)
+
+    title_adjs = [
+      'Beautiful',
+      'Spacious',
+      'Great',
+      'Wonderful',
+      'Best',
+      'Cheap',
+      'Luxurious'
+    ]
+    title_adj_idx = rand(title_adjs.count-1)
+    title_adj = title_adjs[title_adj_idx]
+
+    title_nouns = [
+      'Home',
+      'Room',
+      'Place',
+      'House',
+      'Apartment',
+      'Haven',
+      'Condo',
+      'Residence'
+    ]
+    title_noun_idx = rand(title_nouns.count-1)
+    title_noun = title_nouns[title_noun_idx]
+
+    room_types = [
+      'Entire House/Apt',
+      'Private Room',
+      'Shared Room'
+    ]
+    room_type_idx = rand(3)
+    room_type = room_types[room_type_idx]
+
+    post_code = rand(99999)
+    price_per_night = rand(1000) + 10
+
+    bed_count = rand(10) + 1
+    max_guests = rand(10) + 1
+
+    spot_pic_urls = [
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900936/spot17_qlwvzm.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900954/spot8_hecbj4.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot7_es4acz.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900935/spot18_moq6jj.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot12_rwluji.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900936/spot11_fytnsk.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot14_tdczmw.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot2_ukcmjb.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900939/spot19_x2exoh.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot13_av1dor.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900937/spot4_asyv66.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900940/spot15_ujvbug.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900941/spot9_tfltnp.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900942/spot1_mn0cen.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900956/chicago_sjvjc9.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900943/spot20_bqqzsx.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900961/spot10_mysua7.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900965/spot6_kmsylw.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900958/spot16_dfmr4l.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900956/spot3_ooxov4.jpg",
+      "https://res.cloudinary.com/ddgtwtbre/image/upload/v1478900942/spot5_cwagxm.jpg"
+    ]
+    spot_pic_url_idx = rand(spot_pic_urls.count-1)
+    spot_pic_url = spot_pic_urls[spot_pic_url_idx]
+
+    gen_host_id = rand(67) + 2
+
+    spot.host_id = gen_host_id
+    spot.host_name = 'Host'
+    spot.title = "#{title_adj} #{title_noun} in #{city}"
+    spot.description = "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
+    spot.post_code = "#{post_code}"
+    spot.street_address = "12345 fake street address"
+    spot.price_per_night = price_per_night
+    spot.room_type = "#{room_type}"
+    spot.bed_count = bed_count
+    spot.max_guests = max_guests
+    spot.spot_pic_url = spot_pic_url
+
+    spot.save!
+  end
+end
