+++ db/user_seeds.rb
@@ -1,0 +1,559 @@
+User.create({
+  first_name: "Guest",
+  last_name: "User",
+  email: "guest@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482131647/couchbnb/person-solid_telh7f.png"
+})
+
+User.create({
+  first_name: "Jay",
+  last_name: "Hwang",
+  email: "jay_hwang@gmail.com",
+  password: "password",
+  profile_pic_url: "http://res.cloudinary.com/ddgtwtbre/image/upload/v1481587215/portfolio/linkedin_eiulzq.jpg"
+})
+
+User.create({
+  first_name: "Adam",
+  last_name: "Foosaner",
+  email: 'adam@gmail.com',
+  password: 'password',
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Adam-Foosaner_nypijr.png"
+})
+
+User.create({
+  first_name: "Alex",
+  last_name: "Bennet",
+  email: "alex@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alex-Bennet_bpdgy6.png"
+})
+
+User.create({
+  first_name: "Alex",
+  last_name: "Manning",
+  email: "alex_manning@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alex-Manning_tzlryt.png"
+})
+
+User.create({
+  first_name: "Alex",
+  last_name: "Doundakov",
+  email: "alex_doundakov@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128656/couchbnb/Alexander-Doundakov_g045nm.png"
+})
+
+User.create({
+  first_name: "Amanda",
+  last_name: "Fielding",
+  email: "amanda@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128657/couchbnb/Amanda-Fielding_fnmudx.png"
+})
+
+User.create({
+  first_name: "Andrew",
+  last_name: "Tae",
+  email: "andrew_tae@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Andrew-Tae_ht1kcz.png"
+})
+
+User.create({
+  first_name: "Anudeep",
+  last_name: "Kasturi",
+  email: "anudeep_kasturi@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128664/couchbnb/Anudeep-Kasturi_odxgby.png"
+})
+
+User.create({
+  first_name: "Benjamin",
+  last_name: "Dippelsman",
+  email: "ben_dip@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128658/couchbnb/Benjamin-Dippelsman_xl0tbl.png"
+})
+
+User.create({
+  first_name: "Brian",
+  last_name: "Chen",
+  email: "brian_chen@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/Brian-Chen_mwcxww.png"
+})
+
+User.create({
+  first_name: "Brian",
+  last_name: "Lee",
+  email: "brian_lee@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128666/couchbnb/Brian-Lee_l4b2h5.png"
+})
+
+User.create({
+  first_name: "Brian",
+  last_name: "Tsai",
+  email: "brian_tsai@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Brian-Tsai_n5qaky.png"
+})
+
+User.create({
+  first_name: "Caleb",
+  last_name: "Ontiveros",
+  email: "caleb_onti@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128661/couchbnb/Caleb-Ontiveros_ihuvpj.png"
+})
+
+User.create({
+  first_name: "Christopher",
+  last_name: "Button",
+  email: "christopher_button@gmail.com",
+  password: 'password',
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/Christopher-Button_o3lysh.png"
+})
+
+User.create({
+  first_name: "Cindy",
+  last_name: "Wang",
+  email: "cindy_wang@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/Cindy-Wang_woewj1.png"
+})
+
+User.create({
+  first_name: "Connor",
+  last_name: "Murphy",
+  email: "connor_muphy@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128668/couchbnb/Connor-Murphy_kmoitq.png"
+})
+
+User.create({
+  first_name: "Daniel",
+  last_name: "Sulfaro",
+  email: "daniel_sulfaro@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128667/couchbnb/Daniel-Sulfaro_lnsv0o.png"
+})
+
+User.create({
+  first_name: "David",
+  last_name: "Kim",
+  email: "david_kim@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128665/couchbnb/David-Kim_nqb3b7.png"
+})
+
+User.create({
+  first_name: "David",
+  last_name: "McAfee",
+  email: 'david_mcafee@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/David-McAfee_v2okis.png"
+})
+
+User.create({
+  first_name: "Debra",
+  last_name: "Fong",
+  email: "debra_fong@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128668/couchbnb/Debra-Fong_junc34.png"
+})
+
+User.create({
+  first_name: "Deep",
+  last_name: "Tailor",
+  email: "deep_tailor@gmail.com",
+  password: "password",
+  profile_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/Deep-Tailor_fcid09.png'
+})
+
+User.create({
+  first_name: "Duc",
+  last_name: "Nguyen",
+  email: "duc_nguyen@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128670/couchbnb/Duc-Nguyen_tqn8ll.png"
+})
+
+User.create({
+  first_name: "Eric",
+  last_name: "Kwok",
+  email: "eric_kwok@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128669/couchbnb/Eric-Kwok_tijkgu.png"
+})
+
+User.create({
+  first_name: "Eric",
+  last_name: "Oliveira",
+  email: 'eric_oliveira@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Eric-Oliveira_wpe6od.png"
+})
+
+User.create({
+  first_name: "Ethan",
+  last_name: "Wan",
+  email: "ethan_wan@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Ethan-Wan_ekbgwv.png"
+})
+
+User.create({
+  first_name: "Evan",
+  last_name: "Waits",
+  email: "evan_waits@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Ethan-Wan_ekbgwv.png"
+})
+
+User.create({
+  first_name: "Fan",
+  last_name: "Song",
+  email: "fan_song@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Fan-Song_ul7irn.png"
+})
+
+User.create({
+  first_name: "Frank",
+  last_name: "Bi",
+  email: 'frank_bi@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Frank-Bi_yczosf.png"
+})
+
+User.create({
+  first_name: "Gabriel",
+  last_name: "Lewis",
+  email: "gabriel_lewis@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128673/couchbnb/Gabriel-Lewis_srgc8n.png"
+})
+
+User.create({
+  first_name: "Humza",
+  last_name: "Baig",
+  email: 'humza_baig@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Hamza-Baig_jnvfo1.png"
+})
+
+User.create({
+  first_name: "Henry",
+  last_name: "Hsu",
+  email: "henry_hsu@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128672/couchbnb/Hamza-Baig_jnvfo1.png"
+})
+
+User.create({
+  first_name: "James",
+  last_name: "Conklin",
+  email: "james_conklin",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128678/couchbnb/James-Conklin_p0dvzr.png"
+})
+
+User.create({
+  first_name: "Jane",
+  last_name: "Hauf",
+  email: "jane_hauf@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128675/couchbnb/Jane-Hauf_chq1eb.png"
+})
+
+User.create({
+  first_name: "Janine",
+  last_name: "Love",
+  email: 'janine_love@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128676/couchbnb/Janine-Love_fjhlft.png"
+})
+
+User.create({
+  first_name: 'JC',
+  last_name: 'Balcita',
+  email: "jc_balcita@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128679/couchbnb/JC-Balcita_uhjc9l.png"
+})
+
+User.create({
+  first_name: "John",
+  last_name: "Lekili",
+  email: 'john_lekili@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128678/couchbnb/John-Lekili_hjymxe.png"
+})
+
+User.create({
+  first_name: "Joseph",
+  last_name: "McBride",
+  email: 'joseph_mcbride@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128681/couchbnb/Joseph-McBride_rt9bro.png"
+})
+
+User.create({
+  first_name: "Justice",
+  last_name: "Yen",
+  email: "justice_yen@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128679/couchbnb/Justice-Yen_vaqjbg.png"
+})
+
+User.create({
+  first_name: "Lindsey",
+  last_name: "Whitley",
+  email: 'lindsey_whitley@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Lindsey-Whitley_pwedp2.png"
+})
+
+User.create({
+  first_name: "Louis",
+  last_name: "Cruz",
+  email: "louis_cruz@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Louis-Cruz_bwx4za.png"
+})
+
+User.create({
+  first_name: "Luke",
+  last_name: "Wassink",
+  email: "luke_wassink@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Luke-Wassink_dvzgti.png"
+})
+
+User.create({
+  first_name: "Matt",
+  last_name: "Velker",
+  email: "matt_velker@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128680/couchbnb/Matt-Velker_snyprp.png"
+})
+
+User.create({
+  first_name: "Melissa",
+  last_name: "Lau",
+  email: "melissa_lau@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128683/couchbnb/Melissa-Lau_xrxrbh.png"
+})
+
+User.create({
+  first_name: "Micah",
+  last_name: "Sapitsky",
+  email: "micach_sapitsky@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128684/couchbnb/Micah-Sapitsky_yt84jv.png"
+})
+
+User.create({
+  first_name: "Michael",
+  last_name: "Hirabayashi",
+  email: "michael_hirabayashi@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128682/couchbnb/Michael-Hirabayashi_jd05gl.png"
+})
+
+User.create({
+  first_name: "Michael",
+  last_name: "Loschiavo",
+  email: 'michael_loschiavo@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128688/couchbnb/Michael-Loschiavo_kzxqhl.png"
+})
+
+User.create({
+  first_name: "Michael",
+  last_name: "Pak",
+  email: "michael_pak@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Michael-Pak_dqyesg.png"
+})
+
+User.create({
+  first_name: "Michaelangelo",
+  last_name: "Loreti",
+  email: "miki_loreti@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128685/couchbnb/Michaelangelo-Loreti_td9ell.png"
+})
+
+User.create({
+  first_name: 'Mike',
+  last_name: "Barille",
+  email: 'mike_barille@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128687/couchbnb/Mike-Barille_igrjta.png"
+})
+
+User.create({
+  first_name: "Miki",
+  last_name: "Sparks",
+  email: "miki_sparks@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128687/couchbnb/Miki-Sparks_yttcgo.png"
+})
+
+User.create({
+  first_name: "Nicholas",
+  last_name: "Skoularikis",
+  email: 'nicho@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Nicholas-Skoularikis_p71di2.png"
+})
+
+User.create({
+  first_name: "Paul",
+  last_name: "Cathey",
+  email: "paul_cathey@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128688/couchbnb/Paul-Cathey_uqbi6h.png"
+})
+
+User.create({
+  first_name: "Peik",
+  last_name: "Sia",
+  email: "peik_sia@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Peik-Sia_bqpfzr.png"
+})
+
+User.create({
+  first_name: "Pysith",
+  last_name: "Vanuptikul",
+  email: 'pi_van@gmail.com',
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128689/couchbnb/Pysith_Vanuptikul_eh6roj.png"
+})
+
+User.create({
+  first_name: "Samuel",
+  last_name: "Mak",
+  email: "sam_mak@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128692/couchbnb/Samuel-Mak_e9huh3.png"
+})
+
+User.create({
+  first_name: "Sebastian",
+  last_name: "Cruz",
+  email: "seb_cruz@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128691/couchbnb/Sebastian-Cruz_lm5qde.png"
+})
+
+User.create({
+  first_name: "Shaurya",
+  last_name: "Sinha",
+  email: "shaurya_sinha@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128690/couchbnb/Shaurya-Sinha_fbwey8.png"
+})
+
+User.create({
+  first_name: "Sonik",
+  last_name: "Jang",
+  email: "sonik_jang@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Sonik-Jhang_ajrt0r.png"
+})
+
+User.create({
+  first_name: "Taylor",
+  last_name: "Chestnut",
+  email: "taylor_chestnut",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128691/couchbnb/Taylor-Chestnut_wkwtrx.png"
+})
+
+User.create({
+  first_name: "Todd",
+  last_name: "Nestor",
+  email: "todd_nestor@gmail.com",
+  password: "password",
+  profile_pic_url: 'https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128693/Todd-Nestor_ehdiai.png'
+})
+
+User.create({
+  first_name: "Tonia",
+  last_name: "Hsia",
+  email: "tonia_hsia@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Tonia-Hsia_dxh834.png"
+})
+
+User.create({
+  first_name: "Tony",
+  last_name: "Xie",
+  email: "tony_xie@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Tony-Xie_prap45.png"
+})
+
+User.create({
+  first_name: "Travis",
+  last_name: "Ludlum",
+  email: "travis_ludlum@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128693/couchbnb/Travis-Ludlum_dnjeqg.png"
+})
+
+User.create({
+  first_name: "Trevor",
+  last_name: "Scandalios",
+  email: "trevor_scan@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Trevor-Scandalios_lwkfma.png"
+})
+
+User.create({
+  first_name: "Tyler",
+  last_name: "Fields",
+  email: "tyler_fields@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Trevor-Scandalios_lwkfma.png"
+})
+
+User.create({
+  first_name: "Victor",
+  last_name: "Aw",
+  email: "victor_aw@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128694/couchbnb/Victor-Aw_vesrpm.png"
+})
+
+User.create({
+  first_name: "Vince",
+  last_name: "Sangke",
+  email: "vince_sangke@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128695/couchbnb/Vince-Sangkhae_gotqet.png"
+})
+
+User.create({
+  first_name: "Winston",
+  last_name: "Zhao",
+  email: "winston_zhao@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Winston-Zhao_a5pqvw.png"
+})
+
+User.create({
+  first_name: "Yue",
+  last_name: "Gu",
+  email: "yue_gu@gmail.com",
+  password: "password",
+  profile_pic_url: "https://res.cloudinary.com/ddgtwtbre/image/upload/v1482128697/couchbnb/Yue-Gu_nxn82n.png"
+})
